Deprecated Featureable and changes to Assignable
